delivery-api/requirements.txt
alembic==1.6.5
attrs==21.2.0
beautifulsoup4==4.9.3
blinker==1.4
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
config==0.5.0.post0
Consulta-Correios==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.1
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
platformdirs==2.2.0
pycep-correios==5.0.0
PyJWT==2.1.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
Unidecode==1.2.0
urllib3==1.26.6
Werkzeug==2.0.1
zeep==4.1.0
